Backend-HTC2025/pyproject.toml
[project]
name = "backend-htc2025"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.17.1",
    "asyncpg>=0.30.0",
    "fastapi>=0.121.0",
    "firebase-admin>=7.1.0",
    "google-auth>=2.43.0",
    "greenlet>=3.2.4",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
    "websockets>=15.0.1",
]
